Update : 회원 관련 리팩토링
