+++ src/main/java/com/matdang/seatdang/auth/dto/CustomerSignupDto.java
@@ -1,4 +1,4 @@
-package com.matdang.seatdang.member.dto;
+package com.matdang.seatdang.auth.dto;
 
 import com.matdang.seatdang.member.entity.Gender;
 import lombok.Data;
