+++ src/main/resources/templates/customer/fragments/header.html
@@ -62,7 +62,7 @@
                         <li><a th:href="@{/mypage}">마이페이지</a></li>
                         <li><a th:href="@{/}">항목3</a></li>
                         <li><a th:href="@{/templates/storeowner/main}">점주페이지</a></li>
-                        <li><a th:href="@{/mypage}">마이페이지</a></li>
+                        <li><a th:href="@{/mainmypage}">마이페이지(회원수정)</a></li>
                     </ul>
                 </li>
                 <li class="submenu">
